requirements.txt
agno
openai
python-dotenv
uvicorn
fastapi
pydantic
requests
lancedb
pandas
chonkie[openai]
tiktoken
